requirements.txt
anyio==3.6.2
APScheduler==3.10.1
ariadne==0.19.1
bcrypt==4.0.1
black==23.3.0
blinker==1.6.2
boto3==1.26.142
botocore==1.29.142
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coverage==7.3.1
et-xmlfile==1.1.0
flake8==6.0.0
Flask==2.3.2
graphql-core==3.2.3
graphqlclient==0.2.4
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
importlib-metadata==6.6.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==1.0.0
Naked==0.1.32
nodeenv==1.7.0
numpy==1.24.3
openpyxl==3.1.2
packaging==23.1
pandas==2.0.1
pathspec==0.11.1
platformdirs==3.2.0
prisma==0.8.2
psycopg2==2.9.6
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.7
pyflakes==3.0.1
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
s3transfer==0.6.1
shellescape==3.8.1
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
tomli==2.0.1
tomlkit==0.11.7
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.16
uvicorn==0.21.1
Werkzeug==2.3.3
zipp==3.15.0
